pyproject.toml
[project]
name = "wsjrdp-scripts"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pandas-stubs>=2.3.0.250703",
    "paramiko>=3.5.1",
    "psycopg2>=2.9.10",
    "pyyaml>=6.0.2",
    "ruff>=0.12.7",
    "sepaxml>=2.6.2",
    "sshtunnel>=0.4.0",
    "ty>=0.0.1a16",
    "xlsxwriter>=3.2.5",
]
local = [
    "wsjrdp2027",
]
pytest = [
    "pytest>=8.4.1",
]



[tool.uv]
# paramiko 4 breaks sshtunnel 0.4.0
constraint-dependencies = ["paramiko<4"]
default-groups = [
    "dev",
    "local",
    "pytest",
]

[tool.uv.sources]
wsjrdp2027 = { workspace = true }

[tool.uv.workspace]
members = [
    "packages/wsjrdp2027",
]



[tool.pytest.ini_options]
addopts = "--doctest-modules"
doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"



[tool.ruff]
extend-exclude = ["2023", "2025-05-22--Mailing-UL-Info.py"]
